Update Game.java
Added a key that when the player passes by it, changes the level. This key will be used to allow the player to go to the next level once he reaches a specific kill count. I had to improve certain things such as the way the player interacted with the map, because the previous code was giving me errors, especially with the inclusion of the new maps. I had to create some methods to make the process easier as well.